requirements.txt
absl-py==2.1.0
accelerate==0.32.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
azure-core==1.30.2
azure-identity==1.17.1
azure-storage-blob==12.21.0
bert-score==0.3.13
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
DataProperty==1.0.1
datasets==2.20.0
dill==0.3.8
dnspython==2.6.1
email_validator==2.2.0
evaluate==0.4.2
fastapi==0.111.1
fastapi-cli==0.0.4
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.5.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.0
idna==3.7
intel-openmp==2021.4.0
isodate==0.6.1
Jinja2==3.1.4
joblib==1.4.2
jsonlines==4.0.0
kiwisolver==1.4.5
lm_eval==0.4.3
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mbstrdecoder==1.1.3
mdurl==0.1.2
mkl==2021.4.0
more-itertools==10.3.0
mpmath==1.3.0
msal==1.30.0
msal-extensions==1.2.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.3
nltk==3.8.1
numexpr==2.10.1
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pathvalidate==3.2.0
peft==0.11.1
pillow==10.4.0
pip==23.2.1
portalocker==2.10.1
psutil==6.0.0
pyarrow==17.0.0
pyarrow-hotfix==0.6
pybind11==2.13.1
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
pytablewriter==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
pywin32==306
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rouge_score==0.1.2
sacrebleu==2.4.2
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
setuptools==65.5.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sqlitedict==2.1.0
starlette==0.37.2
sympy==1.13.0
tabledata==1.3.3
tabulate==0.9.0
tbb==2021.13.0
tcolorpy==0.1.6
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.1
tqdm==4.66.4
tqdm-multiprocess==0.0.11
transformers==4.42.4
typepy==1.3.2
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.1
watchfiles==0.22.0
websockets==12.0
word2number==1.1
xxhash==3.4.1
yarl==1.9.4
zstandard==0.23.0
